general cleanup

+++ Software/control.c
@@ -37,27 +37,27 @@
 
 /* Tunable parameters for PI controller */
 /* Gains are Q8 fixed point (actual_gain = GAIN_Q / 256.0) */
-static const int32_t KP_Q = 128; // 0.5 - Proportional gain
-static const int32_t KI_Q = 16;  // 0.0625 - Integral gain
+static const long KP_Q = 128; // 0.5 - Proportional gain
+static const long KI_Q = 16;  // 0.0625 - Integral gain
 
 /*
  * Scale factor to convert PI output (in error units) to DAC counts.
  * Larger values reduce how much DAC moves for a given error.
  */
-static const int32_t PI_TO_DAC_SCALE = 1000; // divide PI output by this to get DAC delta
+static const long PI_TO_DAC_SCALE = 1000; // divide PI output by this to get DAC delta
 
 /*
  * Rate limiting and smoothing
  */
-static const int32_t MAX_STEP = 64; // max DAC counts change per update
+static const long MAX_STEP = 64; // max DAC counts change per update
 
 /*
  * Low-pass filter alpha (Q8). 64 -> 0.25, higher = faster response
  */
-static const int32_t LPF_ALPHA_Q = 64; // 0.25
+static const long LPF_ALPHA_Q = 64; // 0.25
 
-static int32_t integral = 0;
-static uint16_t filtered_output = DAC_MIDPOINT;
+static long integral = 0;
+static unsigned int filtered_output = DAC_MIDPOINT;
 
 /*************************************************************************************/
 /* Control loop functions                                                            */
@@ -65,6 +65,8 @@ static uint16_t filtered_output = DAC_MIDPOINT;
 
 /*
  * Initialize the control module.
+ *
+ * @return     None
  */
 void control_init(void) {
     integral = 0;
@@ -74,45 +76,48 @@ void control_init(void) {
 
 /*
  * Update the control loop with a new error value.
+ *
+ * @param error  The current frequency error (measured - target) in Hz
+ * @return       None
  */
-void control_update(int32_t error) {
+void control_update(long error) {
     // PI controller terms (Q8 math)
-    int32_t p = (KP_Q * error) >> 8;
+    long p = (KP_Q * error) >> 8;
 
     integral += error; // discrete integral (sample period included implicitly)
-    int32_t i = (KI_Q * integral) >> 8;
+    long i = (KI_Q * integral) >> 8;
 
-    int32_t pi_output = p + i; // PI controller output
+    long pi_output = p + i; // PI controller output
 
     // Map PI output to DAC delta
-    int32_t delta = pi_output / PI_TO_DAC_SCALE;
+    long delta = pi_output / PI_TO_DAC_SCALE;
 
     // Apply rate limiting
-    if (delta > (int32_t)MAX_STEP)
-        delta = (int32_t)MAX_STEP;
-    else if (delta < -(int32_t)MAX_STEP)
-        delta = -(int32_t)MAX_STEP;
+    if (delta > MAX_STEP)
+        delta = (long)MAX_STEP;
+    else if (delta < -MAX_STEP)
+        delta = -MAX_STEP;
 
     // Compute raw new value
-    int32_t raw_new = (int32_t)filtered_output + delta;
+    long raw_new = (long)filtered_output + delta;
 
     // Clamp to DAC range
     if (raw_new < 0L)
         raw_new = 0L;
-    if (raw_new > (int32_t)(DAC_RESOLUTION - 1U))
-        raw_new = (int32_t)(DAC_RESOLUTION - 1U);
+    if (raw_new > (long)(DAC_RESOLUTION - 1U))
+        raw_new = (long)(DAC_RESOLUTION - 1U);
 
     // Low-pass filter: filtered = alpha*new + (1-alpha)*old, using Q8
-    int32_t alpha = LPF_ALPHA_Q;
-    int32_t old = (int32_t)filtered_output;
-    int32_t filtered = ((alpha * raw_new) + ((256L - alpha) * old)) >> 8;
+    long alpha = LPF_ALPHA_Q;
+    long old = (long)filtered_output;
+    long filtered = ((alpha * raw_new) + ((256L - alpha) * old)) >> 8;
 
     // Final clamp and write
     if (filtered < 0L)
         filtered = 0L;
-    if (filtered > (int32_t)(DAC_RESOLUTION - 1U))
-        filtered = (int32_t)(DAC_RESOLUTION - 1U);
+    if (filtered > (long)(DAC_RESOLUTION - 1U))
+        filtered = (long)(DAC_RESOLUTION - 1U);
 
-    filtered_output = (uint16_t)filtered;
+    filtered_output = (unsigned int)filtered;
     dac_set_raw(filtered_output);
 }
