+++ Software/date.c
@@ -13,6 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+#include "config.h"
 #include "date.h"
 #include <stdio.h>
 #include <string.h>
@@ -20,6 +21,11 @@
 /*
  * Compute Julian Day Number from Gregorian date
  * Valid for Gregorian calendar dates.
+ * 
+ * @param y  Year (e.g., 2025)
+ * @param m  Month (1-12)
+ * @param d  Day (1-31)
+ * @return   Julian Day Number
  */
 static long jdn_from_ymd(int y, int m, int d)
 {
@@ -30,7 +36,15 @@ static long jdn_from_ymd(int y, int m, int d)
     return jdn;
 }
 
-/* Convert JDN back to y,m,d using Fliegel & Van Flandern algorithm */
+/* 
+ * Convert JDN back to y,m,d using Fliegel & Van Flandern algorithm 
+ * 
+ * @param jdn  Julian Day Number
+ * @param y    Pointer to store year
+ * @param m    Pointer to store month
+ * @param d    Pointer to store day
+ * @return     None
+ */
 static void ymd_from_jdn(long jdn, int *y, int *m, int *d)
 {
     long l = jdn + 68569;
@@ -50,8 +64,13 @@ static void ymd_from_jdn(long jdn, int *y, int *m, int *d)
 
 /*
  * Apply a timezone offset to a UTC date/time.
+ * 
+ * @param utc             Pointer to input UTC date/time
+ * @param out             Pointer to output date/time with offset applied
+ * @param offset_minutes  Timezone offset in minutes (e.g., -300 for UTC-05:00)
+ * @return                None
  */
-void date_apply_offset(const gps_datetime_t *utc, gps_datetime_t *out, int16_t offset_minutes)
+void date_apply_offset(const gps_datetime_t *utc, gps_datetime_t *out, int offset_minutes)
 {
     if (!utc || !out) return;
     if (utc->valid != GPS_VALID) {
@@ -93,6 +112,10 @@ void date_apply_offset(const gps_datetime_t *utc, gps_datetime_t *out, int16_t o
 /**
  * Format time (HH:MM) for display. Buffer should be at least 6 chars long.
  * If invalid, returns "--:--".
+ * 
+ * @param buf  Pointer to output buffer
+ * @param dt   Pointer to gps_datetime_t structure
+ * @return     None
  */
 void date_format_time_short(char *buf, const gps_datetime_t *dt)
 {
@@ -107,8 +130,12 @@ void date_format_time_short(char *buf, const gps_datetime_t *dt)
 /**
  * Create an offset string in the form "+HH:MM" or "-HH:MM" for non-zero offsets.
  * For zero offset returns "+00:00". Buffer should be at least 7 bytes.
+ * 
+ * @param offset_minutes  Timezone offset in minutes
+ * @param buf             Pointer to output buffer
+ * @return                None
  */
-void tz_offset_to_string(int16_t offset_minutes, char *buf)
+void tz_offset_to_string(int offset_minutes, char *buf)
 {
     if (!buf) return;
     char sign_char = (offset_minutes >= 0) ? '+' : '-';
